+++ entrypoints/popup/App.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import { Tabs } from 'wxt/browser';
+import {Tabs} from 'wxt/browser';
 
 // 获取网站地址
 let currentUrl: string;
@@ -17,9 +17,11 @@ onMounted(async () => {
   }
 
   isCSDN.value = currentUrl.includes("https://blog.csdn.net");
-  isZhihu.value = currentUrl.includes("https://zhuanlan.zhihu.com");
+  isZhihuArticle.value = currentUrl.includes("https://zhuanlan.zhihu.com");
+  isZhihuAnswer.value = currentUrl.includes("https://www.zhihu.com/question");
   isCnBlogs.value = currentUrl.includes("https://www.cnblogs.com");
 })
+
 // csdn
 const isCSDN = ref(false);
 const getCSDN = async () => {
@@ -33,14 +35,30 @@ const getCSDN = async () => {
   download(result.title, result.content);
   isLoading.value = false;
 }
-// 知乎
-const isZhihu = ref(false);
-const getZhihu = async () => {
-  isLoaing.value = true;
+
+// 知乎专栏
+const isZhihuArticle = ref(false);
+const getZhihuArticle = async () => {
+  isLoading.value = true;
   // 通信注入脚本
   let content = await browser.scripting.executeScript({
-    target: { tabId },
-    files: ['content-scripts/zhihu.js']
+    target: {tabId},
+    // files: ['content-scripts/zhihu.js']
+    files: ['content-scripts/zhihu_article.js']
+  })
+  const result = content[0].result;
+  download(result.title, result.content);
+  isLoading.value = false;
+}
+
+// 知乎回答
+const isZhihuAnswer = ref(false);
+const getZhihuAnswer = async () => {
+  isLoading.value = true;
+  // 通信注入脚本
+  let content = await browser.scripting.executeScript({
+    target: {tabId},
+    files: ['content-scripts/zhihu_answer.js']
   })
   const result = content[0].result;
   download(result.title, result.content);
@@ -109,8 +127,12 @@ const download = (title: string, content: string) => {
       <button @click="getCSDN" :disabled="isLoading">下载CSDN文章Markdown</button>
       <div class="or">or</div>
     </div>
-    <div class="target" v-show="isZhihu">
-      <button @click="getZhihu" :disabled="isLoaing">下载知乎专栏Markdown</button>
+    <div class="target" v-show="isZhihuArticle">
+      <button @click="getZhihuArticle" :disabled="isLoading">下载知乎专栏Markdown</button>
+      <div class="or">or</div>
+    </div>
+    <div class="target" v-show="isZhihuAnswer">
+      <button @click="getZhihuAnswer" :disabled="isLoading">下载知乎回答Markdown</button>
       <div class="or">or</div>
     </div>
     <div class="target" v-show="isCnBlogs">
